Break sofia into modules
